@@ -2,7 +2,7 @@
 name = "site-slave"
 version = "0.1.0"
 authors = ["IAmTomahawkx <iamtomahawkx@gmail.com>"]
-edition = "2018"
+edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
