@@ -14,72 +14,24 @@ pub async fn img_to_webp(img: &str, target: &String) {
         .expect("failed to remove the tmp file");
 }
 
-pub async fn webp_to_jpg(target: &str, size: (i32, i32)) -> String {
-    let tmp_file = format!("./cache/{}@{}x{}.jpg", target, size.0, size.1);
+
+pub fn webp_to_x(x: &str, target: &str, size: (i32, i32)) -> String {
+    let tmp_file = format!("./cache/{target}@{}x{}.{x}", size.0, size.1);
     if size == (-1, -1) {
         std::process::Command::new("magick")
-            .arg("convert")
-            .arg(format!("./data/{}", target))
-            .arg("jpg:".to_owned() + &tmp_file)
+            .arg(format!("./data/{target}"))
+            .arg(&tmp_file)
             .status()
-            .expect("a");
+            .expect("Failed to convert");
     } else {
         std::process::Command::new("magick")
-            .arg("convert")
-            .arg(format!("./data/{}", target))
+            .arg(format!("./data/{target}"))
+            .arg(&tmp_file)
             .arg("-resize")
             .arg(format!("{}x{}", size.0, size.1))
-            .arg("jpg:".to_owned() + &tmp_file)
             .status()
-            .expect("a");
+            .expect("Failed to convert");
     }
-
-    tmp_file
-}
-
-pub async fn webp_to_png(target: &str, size: (i32, i32)) -> String {
-    let tmp_file = format!("./cache/{}@{}x{}.png", target, size.0, size.1);
-    if size == (-1, -1) {
-        std::process::Command::new("magick")
-            .arg("convert")
-            .arg(format!("./data/{}", target))
-            .arg("png:".to_owned() + &tmp_file)
-            .status()
-            .expect("a");
-    } else {
-        std::process::Command::new("magick")
-            .arg("convert")
-            .arg(format!("./data/{}", target))
-            .arg("-resize")
-            .arg(format!("{}x{}", size.0, size.1))
-            .arg("png:".to_owned() + &tmp_file)
-            .status()
-            .expect("a");
-    }
-
-    tmp_file
-}
-
-pub async fn webp_to_svg(target: &str, size: (i32, i32)) -> String {
-    let tmp_file = format!("./cache/{}@{}x{}.svg", target, size.0, size.1);
-    if size == (-1, -1) {
-        std::process::Command::new("magick")
-            .arg("convert")
-            .arg(format!("./data/{}", target))
-            .arg("svg:".to_owned() + &tmp_file)
-            .status()
-            .expect("a");
-    } else {
-        std::process::Command::new("magick")
-            .arg("convert")
-            .arg(format!("./data/{}", target))
-            .arg("-resize")
-            .arg(format!("{}x{}", size.0, size.1))
-            .arg("svg:".to_owned() + &tmp_file)
-            .status()
-            .expect("a");
-    }
-
     tmp_file
 }
 
@@ -101,10 +53,8 @@ fn as_is(target: &str, size: (i32, i32)) -> String {
 }
 
 pub async fn convert_output(filetype: &str, target: &str, size: (i32, i32)) -> String {
-    return match filetype {
-        "png" => webp_to_png(target, size).await,
-        "jpg" | "jpeg" => webp_to_jpg(target, size).await,
-        "svg" => webp_to_svg(target, size).await,
+    return match filetype.to_lowercase().as_str() {
+        "png" | "jpg" | "jpeg" | "svg" | "heic" => webp_to_x(&filetype, target, size),
         _ => as_is(target, size),
     };
 }
@@ -115,8 +65,8 @@ async fn no_conversion(tmp_file: &str, target: &str) {
 }
 
 pub async fn convert_intake(filetype: &str, tmp_file: &str, target: &str) {
-    return match filetype {
-        "png" | "jpg" | "jpeg" | "svg" => img_to_webp(&tmp_file, &target.to_string()).await,
+    match filetype.to_lowercase().as_str() {
+        "png" | "jpg" | "jpeg" | "svg" | "heic" => img_to_webp(&tmp_file, &target.to_string()).await,
         _ => no_conversion(&tmp_file, &target).await,
     };
 }
