@@ -42,24 +42,24 @@ impl StoredItem {
         pth.exists()
     }
 
-    fn to_named_file_with_size(&self, size: (i32, i32)) -> Result<NamedFile, std::io::Error> {
-        if self.filetype.eq(stored_extension(&self.filetype.to_str())) && self.exists() {
+    async fn to_named_file_with_size(&self, size: (i32, i32)) -> Result<NamedFile, std::io::Error> {
+        if self.filetype.eq(stored_extension(&self.filetype.as_str())) && self.exists() {
             let formatted = format!("data/{:}.{:}", self.name, stored_extension(&self.filetype));
             NamedFile::open(formatted)
         }
         else {
-            let pth = imaging::convert_output(self.filetype, self.name, size);
+            let pth = imaging::convert_output(self.filetype.as_str(), self.name.as_str(), size).await;
             NamedFile::open(pth)
         }
     }
 
-    fn to_named_file(&self) -> Result<NamedFile, std::io::Error> {
-        if self.filetype.eq(stored_extension(&self.filetype.to_str())) && self.exists() {
+    async fn to_named_file(&self) -> Result<NamedFile, std::io::Error> {
+        if self.filetype.eq(stored_extension(&self.filetype.as_str())) && self.exists() {
             let formatted = format!("data/{:}.{:}", self.name, stored_extension(&self.filetype));
             NamedFile::open(formatted)
         }
         else {
-            let pth = imaging::convert_output(self.filetype, self.name, (-1, -1));
+            let pth = imaging::convert_output(self.filetype.as_str(), self.name.as_str(), (-1, -1)).await;
             NamedFile::open(pth)
         }
     }
